services/safety-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiokafka==0.10.0
redis==5.0.4
httpx==0.27.0
numpy==1.26.0
pydantic==2.7.0
python-dotenv==1.0.1
ortools==9.8.3296
